Support :lineno as a line number fragment.
This change allows the common :lineno format to be used to specify a
line number to jump to inside a file. To make the change as small as
possible the format is translated to match the existing line number
fragment specifier line=lineno.

The fragment will only match if the end of the URI is a colon and digits
until the end of the string. This would be a rare case in a most URIs.